agent/requirements.txt
claude-agent-sdk>=0.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
aiosqlite>=0.20.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
